refactor(dataverse): some renaming
