+++ contracts/okp4-dataverse/src/credential/proof.rs
@@ -1,5 +1,5 @@
 use crate::credential::error::InvalidCredentialError;
-use crate::credential::rdf_markers::RDF_TYPE;
+use crate::credential::rdf_marker::RDF_TYPE;
 use itertools::Itertools;
 use okp4_rdf::dataset::{Dataset, QuadIterator};
 use rio_api::model::Term;
